requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Django==2.2.1
django-cors-middleware==1.3.1
django-oauth-toolkit==1.2.0
djangorestframework==3.9.4
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.0.0
pylint==2.3.1
PyMySQL==0.9.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.25.3
wrapt==1.11.1
